refactor project get and list
